+++ test/test_SM_weinberg_angle_meta.cpp
@@ -15,57 +15,6 @@
 using namespace flexiblesusy;
 using namespace weinberg_angle;
 
-void setup_data(const SM_mass_eigenstates& model,
-                Weinberg_angle::Data& data)
-{
-   const double mw_pole = Electroweak_constants::MW;
-   const double mz_pole = Electroweak_constants::MZ;
-   // use mt_drbar instead of mt_pole such that tests work up to high precision
-   const double mt_pole = 165.0;
-   const double scale   = model.get_scale();
-   const double gY      = model.get_g1() * Sqrt(0.6);
-   const double g2      = model.get_g2();
-   const double e_drbar = gY * g2 / Sqrt(Sqr(gY) + Sqr(g2));
-
-   BOOST_CHECK_EQUAL(scale, mz_pole);
-
-   data.fermi_contant  = Electroweak_constants::gfermi;
-   data.mw_pole        = mw_pole;
-   data.mz_pole        = mz_pole;
-   data.mt_pole        = mt_pole;
-   data.scale          = scale;
-   data.gY             = gY;
-   data.g2             = g2;
-   data.g3             = model.get_g3();
-   data.alpha_em_drbar = Sqr(e_drbar) / (4.0 * Pi);
-   data.mh_drbar       = model.get_Mhh();
-
-   const double pizztMZ = Re(model.self_energy_VZ_1loop(mz_pole));
-   const double piwwtMW = Re(model.self_energy_VWp_1loop(mw_pole));
-   const double piwwt0  = Re(model.self_energy_VWp_1loop(0.));
-
-   Weinberg_angle::Self_energy_data se_data;
-   se_data.scale    = scale;
-   se_data.mt_pole  = mt_pole;
-   se_data.mt_drbar = model.get_MFu(2);
-   se_data.mb_drbar = model.get_MFd(2);
-   se_data.gY       = gY;
-   se_data.g2       = g2;
-
-   const double pizztMZ_corrected =
-      Weinberg_angle::replace_mtop_in_self_energy_z(pizztMZ, mz_pole, se_data);
-
-   const double piwwtMW_corrected =
-      Weinberg_angle::replace_mtop_in_self_energy_w(piwwtMW, mw_pole, se_data);
-
-   const double piwwt0_corrected =
-      Weinberg_angle::replace_mtop_in_self_energy_w(piwwt0, 0., se_data);
-
-   data.self_energy_z_at_mz = pizztMZ_corrected;
-   data.self_energy_w_at_mw = piwwtMW_corrected;
-   data.self_energy_w_at_0  = piwwt0_corrected;
-}
-
 BOOST_AUTO_TEST_CASE( test_rho_2 )
 {
    double r;
@@ -93,102 +42,137 @@ BOOST_AUTO_TEST_CASE( test_rho_2 )
 
 BOOST_AUTO_TEST_CASE( test_delta_vb )
 {
-   SM_mass_eigenstates model;
    SM_input_parameters input;
    input.LambdaIN = 0.25;
 
-   setup_SM_const(model, input);
-   model.set_thresholds(2);
-   model.calculate_DRbar_masses();
+   standard_model::Standard_model sm;
+   SM_mass_eigenstates SM;
+
+   setup_SM_const(sm, input);
+   setup_SM_const(SM, input);
+
+   sm.set_thresholds(2);
+   sm.calculate_DRbar_masses();
+   SM.set_thresholds(2);
+   SM.calculate_DRbar_masses();
 
    const double outrho = 1.0, outsin = 0.48;
 
-   Weinberg_angle::Data data;
-   setup_data(model, data);
-   const double delta_vb_1 =
-      Weinberg_angle::calculate_delta_vb(outrho, outsin, data, false);
-
-   SM_weinberg_angle::Sm_parameters sm_parameters;
+   Weinberg_angle::Sm_parameters sm_parameters;
    sm_parameters.fermi_constant = Electroweak_constants::gfermi;
    sm_parameters.mw_pole = Electroweak_constants::MW;
    sm_parameters.mz_pole = Electroweak_constants::MZ;
    sm_parameters.mt_pole = 165.0;
    sm_parameters.alpha_s = 0.1176;
-   SM_weinberg_angle wein(&model, sm_parameters);
-   const double delta_vb_2 = wein.calculate_delta_vb(outrho, outsin);
+   SM_weinberg_angle::Sm_parameters SM_parameters;
+   SM_parameters.fermi_constant = Electroweak_constants::gfermi;
+   SM_parameters.mw_pole = Electroweak_constants::MW;
+   SM_parameters.mz_pole = Electroweak_constants::MZ;
+   SM_parameters.mt_pole = 165.0;
+   SM_parameters.alpha_s = 0.1176;
 
-   BOOST_CHECK_CLOSE_FRACTION(delta_vb_1, delta_vb_2, 1.0e-6);
+   Weinberg_angle wein_sm(&sm, sm_parameters);
+   SM_weinberg_angle wein_SM(&SM, SM_parameters);
+
+   const double delta_vb_sm =
+      wein_sm.calculate_delta_vb(outrho, outsin);
+   const double delta_vb_SM = wein_SM.calculate_delta_vb(outrho, outsin);
+
+   BOOST_CHECK_CLOSE_FRACTION(delta_vb_sm, delta_vb_SM, 1.0e-16);
 }
 
 BOOST_AUTO_TEST_CASE( test_delta_r )
 {
-   SM_mass_eigenstates model;
    SM_input_parameters input;
    input.LambdaIN = 0.25;
 
-   setup_SM_const(model, input);
-   model.set_thresholds(2);
-   model.calculate_DRbar_masses();
+   standard_model::Standard_model sm;
+   SM_mass_eigenstates SM;
+
+   setup_SM_const(sm, input);
+   setup_SM_const(SM, input);
+
+   sm.set_thresholds(2);
+   sm.calculate_DRbar_masses();
+   SM.set_thresholds(2);
+   SM.calculate_DRbar_masses();
 
    const double outrho = 1.0, outsin = 0.48;
 
-   Weinberg_angle::Data data;
-   setup_data(model, data);
-   const double delta_r_1 =
-      Weinberg_angle::calculate_delta_r(outrho, outsin, data, false);
-
-   SM_weinberg_angle::Sm_parameters sm_parameters;
+   Weinberg_angle::Sm_parameters sm_parameters;
    sm_parameters.fermi_constant = Electroweak_constants::gfermi;
    sm_parameters.mw_pole = Electroweak_constants::MW;
    sm_parameters.mz_pole = Electroweak_constants::MZ;
    sm_parameters.mt_pole = 165.0;
    sm_parameters.alpha_s = 0.1176;
-   SM_weinberg_angle wein(&model, sm_parameters);
-   // initialize self-energies
-   wein.pizzt_MZ = wein.calculate_self_energy_VZ(Electroweak_constants::MZ);
-   wein.piwwt_MW = wein.calculate_self_energy_VWp(Electroweak_constants::MW);
-   wein.piwwt_0  = wein.calculate_self_energy_VWp(0.);
-   const double delta_r_2 = wein.calculate_delta_r_hat(outrho, outsin);
+   SM_weinberg_angle::Sm_parameters SM_parameters;
+   SM_parameters.fermi_constant = Electroweak_constants::gfermi;
+   SM_parameters.mw_pole = Electroweak_constants::MW;
+   SM_parameters.mz_pole = Electroweak_constants::MZ;
+   SM_parameters.mt_pole = 165.0;
+   SM_parameters.alpha_s = 0.1176;
 
-   BOOST_CHECK_CLOSE_FRACTION(delta_r_1, delta_r_2, 2.0e-6);
+   Weinberg_angle wein_sm(&sm, sm_parameters);
+   SM_weinberg_angle wein_SM(&SM, SM_parameters);
+
+   // initialize self-energies
+   wein_sm.pizzt_MZ = wein_sm.calculate_self_energy_VZ(Electroweak_constants::MZ);
+   wein_sm.piwwt_MW = wein_sm.calculate_self_energy_VWp(Electroweak_constants::MW);
+   wein_sm.piwwt_0  = wein_sm.calculate_self_energy_VWp(0.);
+   wein_SM.pizzt_MZ = wein_SM.calculate_self_energy_VZ(Electroweak_constants::MZ);
+   wein_SM.piwwt_MW = wein_SM.calculate_self_energy_VWp(Electroweak_constants::MW);
+   wein_SM.piwwt_0  = wein_SM.calculate_self_energy_VWp(0.);
+
+   const double delta_r_sm = wein_sm.calculate_delta_r_hat(outrho, outsin);
+   const double delta_r_SM = wein_SM.calculate_delta_r_hat(outrho, outsin);
+
+   BOOST_CHECK_CLOSE_FRACTION(delta_r_sm, delta_r_SM, 8e-16);
 }
 
 BOOST_AUTO_TEST_CASE( test_sin_theta )
 {
-   SM_mass_eigenstates model;
    SM_input_parameters input;
    input.LambdaIN = 0.25;
 
-   setup_SM_const(model, input);
-   model.set_thresholds(2);
-   model.calculate_DRbar_masses();
+   standard_model::Standard_model sm;
+   SM_mass_eigenstates SM;
 
-   const double tol = 1.0e-10;
-   const int maxTries = 20;
-   const double rho_start = 1.0, sin_start = 0.48;
+   setup_SM_const(sm, input);
+   setup_SM_const(SM, input);
 
-   Weinberg_angle::Data data;
-   setup_data(model, data);
-   Weinberg_angle wein;
-   wein.disable_susy_contributions();
-   wein.set_data(data);
-   wein.set_number_of_iterations(maxTries);
-   wein.set_precision_goal(tol);
-   const int error = wein.calculate(rho_start, sin_start);
-   const double sin_theta_1 = wein.get_sin_theta();
-   BOOST_REQUIRE(error == 0);
+   sm.set_thresholds(2);
+   sm.calculate_DRbar_masses();
+   SM.set_thresholds(2);
+   SM.calculate_DRbar_masses();
 
-   SM_weinberg_angle::Sm_parameters sm_parameters;
+   static constexpr double tol = 1.0e-10;
+   static constexpr int maxTries = 20;
+   static constexpr double rho_start = 1.0;
+   static constexpr double sin_start = 0.48;
+
+   Weinberg_angle::Sm_parameters sm_parameters;
    sm_parameters.fermi_constant = Electroweak_constants::gfermi;
    sm_parameters.mw_pole = Electroweak_constants::MW;
    sm_parameters.mz_pole = Electroweak_constants::MZ;
    sm_parameters.mt_pole = 165.0;
    sm_parameters.alpha_s = 0.1176;
-   SM_weinberg_angle pwein(&model, sm_parameters);
-   pwein.set_number_of_iterations(maxTries);
-   pwein.set_precision_goal(tol);
-   double sin_theta_2;
-   BOOST_REQUIRE_NO_THROW(sin_theta_2 = pwein.calculate(sin_start).first);
+   Weinberg_angle wein_sm(&sm, sm_parameters);
+   wein_sm.set_number_of_iterations(maxTries);
+   wein_sm.set_precision_goal(tol);
+   double sin_theta_sm;
+   BOOST_REQUIRE_NO_THROW(sin_theta_sm = wein_sm.calculate_sin_theta_w(sin_start));
 
-   BOOST_CHECK_CLOSE_FRACTION(sin_theta_1, sin_theta_2, 6.0e-9);
+   SM_weinberg_angle::Sm_parameters SM_parameters;
+   SM_parameters.fermi_constant = Electroweak_constants::gfermi;
+   SM_parameters.mw_pole = Electroweak_constants::MW;
+   SM_parameters.mz_pole = Electroweak_constants::MZ;
+   SM_parameters.mt_pole = 165.0;
+   SM_parameters.alpha_s = 0.1176;
+   SM_weinberg_angle wein_SM(&SM, SM_parameters);
+   wein_SM.set_number_of_iterations(maxTries);
+   wein_SM.set_precision_goal(tol);
+   double sin_theta_SM;
+   BOOST_REQUIRE_NO_THROW(sin_theta_SM = wein_SM.calculate(sin_start).first);
+
+   BOOST_CHECK_CLOSE_FRACTION(sin_theta_sm, sin_theta_SM, 1.0e-16);
 }